refactor(spiffe-client): move runtime api classes and introduce svid-specific clients
- moving x509 / jwt svids into their packages
- introducing svid-specific client interfaces as typically they will do one or another (x509 or jwt)
- caveat is that https://github.com/quarkusio/quarkus/blob/main/adr/0009-extension-structure.adoc requires probably using `quarkus.spiffe.client.*` subpackages and I moved SVIDs to `quarkus.spiffe.svid.`, I am open to opinions and changes
