+++ extensions/oidc-common/runtime/src/test/java/io/quarkus/oidc/common/runtime/SpiffeClientAssertionProviderTest.java
@@ -15,8 +15,8 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 import io.quarkus.spiffe.client.SpiffeClient;
-import io.quarkus.spiffe.client.WorkloadCertificateDocument;
-import io.quarkus.spiffe.client.WorkloadJsonWebToken;
+import io.quarkus.spiffe.svid.jwt.WorkloadJsonWebToken;
+import io.quarkus.spiffe.svid.x509.WorkloadCertificateDocument;
 import io.smallrye.jwt.build.Jwt;
 import io.smallrye.mutiny.Uni;
 import io.vertx.core.Vertx;
