+++ extensions/spiffe-client/deployment/src/test/java/io/quarkus/spiffe/client/deployment/test/AbstractSpiffeClientTest.java
@@ -31,8 +31,9 @@ import org.junit.jupiter.api.Test;
 import io.quarkus.spiffe.client.SpiffeAuthorizationException;
 import io.quarkus.spiffe.client.SpiffeClient;
 import io.quarkus.spiffe.client.SpiffeConnectionException;
-import io.quarkus.spiffe.client.WorkloadJsonWebToken;
 import io.quarkus.spiffe.client.deployment.SpiffeDevServicesProcessor.ServerMode;
+import io.quarkus.spiffe.svid.jwt.JwtSvidClient;
+import io.quarkus.spiffe.svid.jwt.WorkloadJsonWebToken;
 import io.vertx.core.Vertx;
 import io.vertx.core.buffer.Buffer;
 import io.vertx.core.http.HttpMethod;
@@ -50,6 +51,9 @@ abstract class AbstractSpiffeClientTest {
 
     @Inject
     SpiffeClient spiffeClient;
+
+    @Inject
+    JwtSvidClient jwtSvidClient;
 
     @ConfigProperty(name = BASE_URL_CONFIG_KEY)
     String devServerBaseUrl;
@@ -189,7 +193,7 @@ abstract class AbstractSpiffeClientTest {
     }
 
     private WorkloadJsonWebToken fetchSvid(String audience) {
-        return spiffeClient.getWorkloadJsonWebToken(audience).await().atMost(TIMEOUT);
+        return jwtSvidClient.getWorkloadJsonWebToken(audience).await().atMost(TIMEOUT);
     }
 
     private void setServerMode(ServerMode mode) {
