+++ extensions/spiffe-client/runtime-api/src/main/java/io/quarkus/spiffe/svid/x509/WorkloadCertificateDocument.java
@@ -1,4 +1,4 @@
-package io.quarkus.spiffe.client;
+package io.quarkus.spiffe.svid.x509;
 
 import io.smallrye.common.annotation.Experimental;
