+++ extensions/spiffe-client/runtime/src/main/java/io/quarkus/spiffe/client/runtime/internal/SpiffeClientImpl.java
@@ -26,16 +26,16 @@ import io.quarkus.runtime.configuration.ConfigurationException;
 import io.quarkus.spiffe.client.SpiffeAuthorizationException;
 import io.quarkus.spiffe.client.SpiffeClient;
 import io.quarkus.spiffe.client.SpiffeConnectionException;
-import io.quarkus.spiffe.client.WorkloadCertificateChain;
-import io.quarkus.spiffe.client.WorkloadCertificateDocument;
-import io.quarkus.spiffe.client.WorkloadJsonWebToken;
-import io.quarkus.spiffe.client.WorkloadTrustBundle;
 import io.quarkus.spiffe.client.runtime.internal.proto.JWTSVID;
 import io.quarkus.spiffe.client.runtime.internal.proto.JWTSVIDRequest;
 import io.quarkus.spiffe.client.runtime.internal.proto.JWTSVIDResponse;
 import io.quarkus.spiffe.client.runtime.internal.proto.X509SVID;
 import io.quarkus.spiffe.client.runtime.internal.proto.X509SVIDRequest;
 import io.quarkus.spiffe.client.runtime.internal.proto.X509SVIDResponse;
+import io.quarkus.spiffe.svid.jwt.WorkloadJsonWebToken;
+import io.quarkus.spiffe.svid.x509.WorkloadCertificateChain;
+import io.quarkus.spiffe.svid.x509.WorkloadCertificateDocument;
+import io.quarkus.spiffe.svid.x509.WorkloadTrustBundle;
 import io.smallrye.common.os.OS;
 import io.smallrye.mutiny.Multi;
 import io.smallrye.mutiny.Uni;
